+++ db-collections/PermissionedDataPersister.ts
@@ -2,13 +2,13 @@
 /** {@link DataPersister.Adapter} wrapper that checks permissions before reading/writing data
  * @author TeamworkGuy2
  */
-class PermissionedDataPersisterAdapter implements DataPersister.Adapter {
-    private persister: DataPersister.Adapter;
+class PermissionedDataPersisterAdapter implements DataPersister {
+    private persister: DataPersister;
     private syncSettings: ReadWritePermission;
     private storeSettings: StorageFormatSettings;
 
 
-    constructor(persister: DataPersister.Adapter, syncSettings: ReadWritePermission, storeSettings: StorageFormatSettings) {
+    constructor(persister: DataPersister, syncSettings: ReadWritePermission, storeSettings: StorageFormatSettings) {
         this.syncSettings = syncSettings;
         this.storeSettings = storeSettings;
         this.persister = persister;
@@ -21,7 +21,7 @@ class PermissionedDataPersisterAdapter implements DataPersister.Adapter {
         }
         else {
             var dfd = Q.defer<any>();
-            dfd.reject("permission denied: data persister write permission denied due to settings");
+            dfd.reject("permission denied: data persister write permissions are denied");
             return dfd.promise;
         }
     }
@@ -33,19 +33,19 @@ class PermissionedDataPersisterAdapter implements DataPersister.Adapter {
         }
         else {
             var dfd = Q.defer<any>();
-            dfd.reject("permission denied: data persister read permission denied due to settings");
+            dfd.reject("permission denied: data persister read permissions are denied");
             return dfd.promise;
         }
     }
 
 
-    public clearPersistenceDb(): Q.Promise<void> {
+    public clearPersistentDb(): Q.Promise<void> {
         if (this.syncSettings.writeAllow) {
-            return this.persister.clearPersistenceDb();
+            return this.persister.clearPersistentDb();
         }
         else {
             var dfd = Q.defer<any>();
-            dfd.reject("permission denied: data persister write permission denied due to settings");
+            dfd.reject("permission denied: data persister write permissions are denied");
             return dfd.promise;
         }
     }